test: add test coverage to fs.promises.mkdir
